requirements.txt
python-dotenv==1.1.0
pywin32==310
